refactor: ToiletAccessibility 큐레이션 → Toilet 큐레이션 페이지 재설계
- scc-api submodule 업데이트(d8c0919) + pnpm codegen
- /toilet-accessibilities → /toilets 페이지 rename
- sourceType 필터/필드 제거, isSearchable 필터만 유지
- AdminToiletDto 기반: hasExternalSource 뱃지, toiletReviewDetails 배열 표시
- useToilets / setToiletSearchable 훅으로 교체 (adminSearchToilets, setToiletSearchable operation)
- 사이드바 메뉴 경로 /toilets로 변경
