+++ app/constants/menu.ts
@@ -48,7 +48,7 @@ export const menuItems: MenuItem[] = [
   { title: "저장 리스트 관리", url: "/placeList", icon: Bookmark },
   { title: "접근성 신고 관리", url: "/accessibilityReport", icon: AlertTriangle },
   { title: "접근성 정보 채우기", url: "/accessibility-suggestion", icon: ScanEye },
-  { title: "장애인 화장실 큐레이션", url: "/toilet-accessibilities", icon: Toilet },
+  { title: "장애인 화장실 큐레이션", url: "/toilets", icon: Toilet },
   { title: "실험 배정 관리", url: "/experiment", icon: FlaskConical },
   { title: "출석부 웹사이트", url: "/attendance", icon: NotepadText },
 ]
